# docs/openapi.yaml
openapi: 3.0.1
info:
  title: L0 Storage
  description: Long term immutable block storage.
  license:
    name: MIT
    url: https://github.com/tiki/l0-storage/blob/main/LICENSE
  version: 0.0.1
externalDocs:
  description: javadocs
  url: coming soon
servers:
- url: coming soon
paths:
  /api/latest/policy:
    post:
      tags:
      - POLICY
      summary: Request a new policy
      operationId: postNew
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyAOReq'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAORsp'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "417":
          description: Expectation Failed
  /api/latest/api-id/new:
    post:
      tags:
      - API ID
      summary: Request a new API Id
      operationId: postNew_1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiIdAORsp'
  /health:
    get:
      tags:
      - HEALTH
      operationId: get
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: OK
  /api/latest/api-id:
    get:
      tags:
      - API ID
      summary: Get all provisioned API Ids
      operationId: getAll
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiPageApiIdAORsp'
  /api/latest/api-id/key/{api-id}:
    get:
      tags:
      - API ID
      summary: Get an API Id's properties
      operationId: getKey
      parameters:
      - name: api-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiIdAORsp'
        "404":
          description: Not Found
    delete:
      tags:
      - API ID
      summary: Revoke an API Id (permanent)
      operationId: deleteKey
      parameters:
      - name: api-id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiIdAORsp'
        "404":
          description: Not Found
components:
  schemas:
    PolicyAORsp:
      type: object
      properties:
        policy:
          type: string
        policySignature:
          type: string
        policyDate:
          type: string
        urnPrefix:
          type: string
        expiresIn:
          type: integer
          format: int64
    PolicyAOReq:
      type: object
      properties:
        pubKey:
          type: string
        signature:
          type: string
        stringToSign:
          type: string
    ApiIdAORsp:
      type: object
      properties:
        apiId:
          type: string
        valid:
          type: boolean
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        detail:
          type: string
        help:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
    ApiPageApiIdAORsp:
      type: object
      properties:
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ApiIdAORsp'
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        page:
          type: integer
          format: int32